Use telia-oss/githubapp package
